Final draft

+++ deploy.sh
@@ -2,22 +2,24 @@
 # update minimal theme:
 #   git submodule update --remote themes/minimal
 
-if [ $# -ne 1 ]; then
-	echo "[ERROR] Please pass all.inkl user ID"; exit 1
+if [ ! -f .deploy ]; then
+	echo "[ERROR] Please provide .deploy file holding USER and DOMAIN"; exit 1
+	exit 1
 fi
 
-webroot="/www/htdocs/$1/boddenberg.it/"
+# load config
+. ./.deploy
+webroot="/www/htdocs/$USER/$DOMAIN/"
 
 echo "[INFO] invoking hugo"
 hugo
 
 echo "[INFO] Deleting current remote web content"
-ssh boddenberg.it mv "${webroot}/usage" "/${webroot}/../usage-boddenberg-it"
+ssh boddenberg.it mv "${webroot}/usage" "/${webroot}/../usage-$DOMAIN"
 ssh boddenberg.it rm -rf "${webroot}*"
-ssh boddenberg.it mv "/${webroot}/../usage-boddenberg-it" "${webroot}/usage"
+ssh boddenberg.it mv "/${webroot}/../usage-$DOMAIN" "${webroot}/usage"
 
 echo "[INFO] Uploading local web content"
 scp -r public/* boddenberg.it:"$webroot"
 
-echo
 echo "[INFO] done."
